Fixing MouseEvent.stop()
The MouseEvent wrapper was calling stop() on its own jso rather than the
original Google Maps event object. It now retains a reference to the
event object so that it can call the method as required.

If we're retaining a reference to the event object maybe we don't need
to hold a separate reference to the event coordinates...